Addresses iPad crashing
