+++ CoffApp.xcodeproj/xcshareddata/xcschemes/CoffUp (watchOS) WatchKit App (Notification).xcscheme
@@ -56,8 +56,10 @@
       allowLocationSimulation = "YES"
       launchAutomaticallySubstyle = "8"
       notificationPayloadFile = "CoffUp (watchOS) WatchKit Extension/PushNotificationPayload.apns">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <RemoteRunnable
+         runnableDebuggingMode = "2"
+         BundleIdentifier = "com.apple.Carousel"
+         RemotePath = "/The Coffee">
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "0FA09F122661FE5700E04777"
@@ -65,7 +67,7 @@
             BlueprintName = "CoffUp (watchOS) WatchKit App"
             ReferencedContainer = "container:CoffApp.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </RemoteRunnable>
    </LaunchAction>
    <ProfileAction
       buildConfiguration = "Release"
@@ -75,8 +77,10 @@
       debugDocumentVersioning = "YES"
       launchAutomaticallySubstyle = "8"
       notificationPayloadFile = "CoffUp (watchOS) WatchKit Extension/PushNotificationPayload.apns">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <RemoteRunnable
+         runnableDebuggingMode = "2"
+         BundleIdentifier = "com.apple.Carousel"
+         RemotePath = "/The Coffee">
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "0FA09F122661FE5700E04777"
@@ -84,7 +88,16 @@
             BlueprintName = "CoffUp (watchOS) WatchKit App"
             ReferencedContainer = "container:CoffApp.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </RemoteRunnable>
+      <MacroExpansion>
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "0FA09F122661FE5700E04777"
+            BuildableName = "CoffUp (watchOS) WatchKit App.app"
+            BlueprintName = "CoffUp (watchOS) WatchKit App"
+            ReferencedContainer = "container:CoffApp.xcodeproj">
+         </BuildableReference>
+      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
