+++ CoffApp.xcodeproj/xcshareddata/xcschemes/CoffUp (watchOS) WatchKit App.xcscheme
@@ -75,8 +75,10 @@
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES"
       notificationPayloadFile = "CoffUp (watchOS) WatchKit Extension/PushNotificationPayload.apns">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <RemoteRunnable
+         runnableDebuggingMode = "2"
+         BundleIdentifier = "com.apple.Carousel"
+         RemotePath = "/The Coffee">
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "0FA09F122661FE5700E04777"
@@ -84,7 +86,7 @@
             BlueprintName = "CoffUp (watchOS) WatchKit App"
             ReferencedContainer = "container:CoffApp.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </RemoteRunnable>
    </LaunchAction>
    <ProfileAction
       buildConfiguration = "Release"
@@ -92,8 +94,10 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <RemoteRunnable
+         runnableDebuggingMode = "2"
+         BundleIdentifier = "com.apple.Carousel"
+         RemotePath = "/The Coffee">
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "0FA09F122661FE5700E04777"
@@ -101,7 +105,16 @@
             BlueprintName = "CoffUp (watchOS) WatchKit App"
             ReferencedContainer = "container:CoffApp.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </RemoteRunnable>
+      <MacroExpansion>
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "0FA09F122661FE5700E04777"
+            BuildableName = "CoffUp (watchOS) WatchKit App.app"
+            BlueprintName = "CoffUp (watchOS) WatchKit App"
+            ReferencedContainer = "container:CoffApp.xcodeproj">
+         </BuildableReference>
+      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
